Remove public functions from ParameterizedTest
This makes `import ParameterizedTest` give you *only* the `param_test` macro, rather than (accidentally) pulling in the `parse_examples` function which is really an implementation detail.

Resolves #19
