+++ lib/parameterized_test/sigil_v115.exs
@@ -68,6 +68,6 @@ defmodule ParameterizedTest.Sigil do
   @spec sigil_PARAMS(String.t(), Keyword.t()) :: [map()]
   # credo:disable-for-next-line Credo.Check.Readability.FunctionNames
   def sigil_PARAMS(table, _opts \\ []) do
-    ParameterizedTest.parse_examples(table)
+    ParameterizedTest.Parser.parse_examples(table)
   end
 end
