+++ test/parameterized_test_test.exs
@@ -268,7 +268,7 @@ defmodule ParameterizedTestTest do
     end
   end
 
-  @module_examples ParameterizedTest.parse_examples("""
+  @module_examples ParameterizedTest.Parser.parse_examples("""
                    | int_1 | int_2 |
                    | 99    | 100   |
                    """)
@@ -332,11 +332,4 @@ defmodule ParameterizedTestTest do
     end
   end
 
-  describe "parse_examples/1" do
-    test "accepts strings that parse as empty" do
-      for empty <- ["", " ", "\n", "\t", "\r\n", "\n\n \r\n \t \r"] do
-        assert ParameterizedTest.parse_examples(empty) == []
-      end
-    end
-  end
 end
